updated changes to pull request for backend transcription feature
