+++ liquid/rekognition.py
@@ -1,10 +1,7 @@
 from typing import Callable
-import boto3
+import boto3, random, string, json
 from flask import current_app as app
 
-import webvtt, string, json
-from datetime import datetime as dt
-from datetime import timezone
 
 from . import s3
 
@@ -105,9 +102,14 @@ class VideoSubmitter:
 
     def _do_text_transcription(self):
         """Calls AWS Rekognition to create captions for the video."""
+
+        def random_char_sequence(y):
+            """Creates a random char sequence of length y for the job_id name randomization"""
+            return "".join(random.choice(string.ascii_letters) for x in range(y))
+
         video_uri = s3.get_object_url(self.video)
-        job_id = f"transcription-service-{self.liquid.id}-0382734283"
-        response = aws_trs.start_transcription_job(
+        job_id = f"transcription-service-{self.liquid.id}-" + random_char_sequence(24)
+        aws_trs.start_transcription_job(
             TranscriptionJobName=job_id,
             Media={"MediaFileUri": video_uri},
             MediaFormat="mp4",
@@ -195,6 +197,8 @@ class VideoDetector:
             "FAILED",
         ]:
             return False
+
+        # Returns the URL of the .vtt caption file
         self.data = response["TranscriptionJob"]["Subtitles"]["SubtitleFileUris"][0]
         return True
